backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
beanie
motor
passlib[bcrypt]
python-jose[cryptography]
email-validator
